CLUSTER-9. bento/bento-env.sh do not work if invoked via symlinks.
Similar to SCHEMA-43, the bento scripts do not work correctly
if they are invoked via symlinks. They will not use the dirname of $0
correctly, so we need to traverse symlinks "manually".

The location of the bento-env.sh script (or symlink) is also tied to the
'build/' directory that the tasktracker puts its state under. This is
now moved under the 'state/' directory regardless of where you invoke
it from.

Signed-off-by: Chad Metcalf <chad@wibidata.com>