CLUSTER-7. bento-cluster will not start if there are DNS problems with hostname and localhost.
This commit modifies bento-cluster to include a tool, DNSCheckingTool, that
accepts the result of running `hostname` as an argument. It then ensures
the following:

  1. That 'localhost' can be resolved to an IP address.
  2. That the passed hostname can be resolved to an IP address.
  3. That the addresses resolved match.

If any of these conditions are not met, bento-cluster will refuse to start.
This is to avoid errors bento-cluster encounters when /etc/hosts is misconfigured,
in particular when localhost and the machine's hostname map to different local
IP addresses (which is the default setup on fresh Ubuntu machines).

Signed-off-by: Lee Sheng <lsheng@wibidata.com>