CLUSTER-19. CLUSTER-7 breaks bento-cluster on Mac OS X.
The DNS checker tool (which ensures both localhost and the machine's hostname
resolve to the same IP and refuses to start bento-cluster if this is not the
case) does not function correctly on Mac OS X. This check is designed to avoid
a common problem on Ubuntu machines. This commit modifies the bento script so
that the DNS check is not performed on Mac OS X, so that bento-cluster again
functions correctly on those systems.

Signed-off-by: Chad Metcalf <chad@wibidata.com>