BENTO-15 (Part 2). bento script compatability with upgrade check-in.
This commit modifies the bento script such that it runs the script
upgrade-daemon-start.sh when bento-cluster starts, and upgrade-daemon-stop.sh
when bento-cluster stops. A script that handles generating a UUID for the user
(generate-uuid.sh) is also hooked in.  This is only done if the scripts are
present, indicating we are running in a Kiji BentoBox.

Signed-off-by: Aaron Kimball <aaron@wibidata.com>